Tweak kiosk Context so we get the experience name
